// DXC_Frontend/src/api/models/supportGroupDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface SupportGroupDto {
  publicId?: string
  categoryPublicId?: string
  /** @nullable */
  categoryName: string | null
  /** @nullable */
  groupName: string | null
  /** @nullable */
  groupLink: string | null
  /** @nullable */
  groupType: string | null
  /** @nullable */
  description?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}